requirements.txt
streamlit>=1.30.0
pillow>=10.0.0
moviepy>=2.0.0
edge-tts>=6.1.0
google-genai>=0.1.0
numpy>=1.24.0
proglog>=0.1.10
openai>=1.0.0